Update to build on 10.9

+++ Source/UKBorderlessWindow.m
@@ -12,7 +12,7 @@
 @implementation UKBorderlessWindow
 
 // Designated Initializer:
--(id)   initWithContentRect: (NSRect)box styleMask: (unsigned int)sty backing: (NSBackingStoreType)bs defer: (BOOL)def
+-(id)   initWithContentRect: (NSRect)box styleMask: (unsigned long)sty backing: (NSBackingStoreType)bs defer: (BOOL)def
 {
 	self = [super initWithContentRect:box styleMask: NSBorderlessWindowMask backing:bs defer: def];
